Made use of the existing timer and medication classes to populate the timer list

+++ app/src/main/java/com/lovelace_scd/timed/io/MedJsonReader.kt
@@ -68,7 +68,7 @@ class MedJsonReader(val filename: String = "src/main/java/com/lovelace_scd/timed
                         tim["doseUnit"] as String,
                         tim["isRefillable"] as Boolean,
                 );
-                timers.add(Timer(med, (tim["baseDate"] as Int).toLong(), tim["skipNextDose"] as Boolean, tim["nextDoseReady"] as Boolean));
+                timers.add(Timer(med, (tim["baseDate"] as Long), tim["skipNextDose"] as Boolean, tim["nextDoseReady"] as Boolean));
             }
         }
         return timers;
